backend/requirements.txt
fastapi[all]>=0.68.0
uvicorn[standard]>=0.15.0

asyncpg>=0.27.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.8.1

pydantic-settings>=2.0.0

celery[redis]>=5.2.7
kombu>=5.2.4

aioredis>=2.0.1

pyjwt>=2.4.0
passlib[bcrypt]>=1.7.4

loguru>=0.6.0

pytest>=7.1.2
httpx>=0.23.0
pytest-asyncio>=0.20.3
